Create new 'createPartialStylelintResult' module
In a future PR this module will be required by the new `browser.js` module. This change has been extracted from: https://github.com/stylelint/stylelint/pull/4796
